adjust consolidations tests for EL consolidations
