+++ tests/generators/operations/main.py
@@ -45,7 +45,7 @@ if __name__ == "__main__":
 
     _new_electra_mods = {key: 'eth2spec.test.electra.block_processing.test_process_' + key for key in [
         'attestation',
-        'consolidation',
+        'execution_layer_consolidation',
         'deposit_receipt',
         'execution_layer_withdrawal_request',
         'voluntary_exit'
